In friends branch [Using python 3.12.7 and CustonTKinter 5.2.2.].
Added two new windows, one to display the current friends and a second to add a new friend.
The display friends window is called from the main menu, there is an Add button to add a friend.
The friends are saved in a csv file [for the moment] in the data directory.
No validation is done on the data entry - will add as needed.

Need to add the edit and delete buttons to the friends window.
Need to test before merging into main.

Also added a sperate version file in the data directory.
	This allows the version in an installed Klock to be updated without uninstalling first.
	The install process doesn't upgrade the existing config file.

Note : in addition to installing the files in the requirements files, the following source
       directories need to be added, see Help [Credits Section] for GitHub addresses.
       CTkDatePicker, CTkPDFViewer & CTkXYFrame