in vfd branch.
Found Vacuum Fluorescent Displays for tk when looking for other things.
	See Help [Credits Section] for location and original author.
	Added config options for vfdKlock, screen location and colour only used.
This is used in the first of the alternative Klocks.
When this Klock is displayed, the main Klock is minimised to the task bar.

Changed the status bar [again] to try and fix the layout problems.
The status bar is now one long label and not three as before.
This makes the logic and settings a lot easier - will see.

Corrected a bug in the screen position locator of the main window.

Added a new time type - getDigitTime - returns current time as a list [H,H,M,M,S,S]
	Used in vfdKlock