TextClock seems to be working, after some tweaking.
Amended to main update method of klock only update the main menu and time in main screen is running.
Apparently, to make a python method private you prefix it with two underscores [__].
I had only been using one, so took the opportunity of the branch to change all the ones to twos.
Also, it appears the file showmenuButtons is no longer used - removed.